common/dir.c: do not include openbox/ dirs
It is better to be explicit with config files and only using those
specifically intended for labwc. Whilst compatibility with openbox is
intended, there are areas where it just doesn't make sense to include
settings intended for openbox, for example in the case of
$HOME/.config/openbox/autostart

With built-in icons and theme/config settings, it should be easy for new
users to get started without spending time on config files anyway.