cursor: Prevent application crash on drag-and-drop
Wlroots resets the focused surface on Drag and Drop.

Before this patch we tried to enforce a client side cursor
image update by forcefully clearing the focus which, even though
`wlr_seat->pointer_state.focused_surface` was `NULL`, was sent to the
drag source. After the clear we are sending an enter event for the
surface under the cursor and if this surface was the drag source itself
it was heavily confused. This caused all kinds of weird behavior from
spawning new drag icons constantly to outright crashing the application.

Reported-by: @Flrian
Bisected-by: @Flrian