cursor: fix layer subsurface focus bug
...by checking keyboard-interactivity before giving layer-subsurface focus
on button press.

Before this patch, waybar's raise-minimize did not work because keyboard
focus was given to waybar layer-shell subsurfaces without checking for
their keyboard interactivity setting.  Clicking on a waybar task-button
gave keyboard focus to that button which removed it from the client and
therefore confused the panel logic.

Fixes: issue #1131